Price column fix

+++ app.py
@@ -9,7 +9,7 @@ st.write('filter the data below to see the vehicles listed')
 
 
 df = pd.read_csv('vehicles_us.csv')
-df['price'] = pd.to_numeric(df['price'], errors='coerce') 
+df['price'] = pd.to_numeric(df['price'], errors='coerce').astype('float64')
 df['price'] = df['price'].fillna(0)
